requirements.txt
numpy==1.26.4
pandas
tqdm
ir_datasets
sentence-transformers
faiss-cpu
streamlit
transformers
torch